requirements.txt
azure-cognitiveservices-speech
python-dotenv
ffmpeg-python
winsound